pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nemotron"
version = "0.1.0"
description = "A coding agent powered by Structural Memory Protocol (SMP)"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "msgspec>=0.19",
    "watchfiles>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["nemotron"]

[project.scripts]
nemotron = "nemotron.main:main"

[tool.ruff]
target-version = "py311"
line-length = 120
